requirements.txt
streamlit
torch
torchvision
torchaudio
numpy
Pillow
opencv-python-headless
matplotlib
transformers
tqdm
scikit-learn